requirements.txt
customtkinter>=5.2.0
Pillow>=10.0.0
openpyxl>=3.1.0
pdfplumber>=0.10.0
python-docx>=1.0.0
pyinstaller>=6.0.0
